feat(User Registration): Users can receive descriptive messages
This commit delivers a feature for users to recieve descriptive
messages on succesfull registration and login.
The user names should be between 3 to 10 characters, user names
should only contain letters, numbers and underscore.
Passwords should be between 8 to 128 characters, a password
should contain atleast on capital letter and a number
Email addresses should be valid.

Tests for the feature are included

[Delivers #162163263]